requirements.txt
streamlit==1.39.0
langchain==0.3.7
langchain-openai==0.1.6
langchain-community
openai==1.57.0
tiktoken==0.8.0
beautifulsoup4==4.12.3
requests==2.32.3
python-dotenv==1.0.0
rank-bm25
